requirements.txt
opencv-python==4.0.0.21
gym==0.15.3
tensorflow-gpu==1.14.0
stable_baselines==2.4.0
# conda install cudnn==7.6.0 -c conda-forge
# conda install cudatoolkit==10.0.130